requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
gigachat==0.1.17
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
